Avoid creating fresh root for the first test in `InstancePerLeaf` (#5201)
<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->

currently, `InstancePerLeaf` creates a fresh root for the first test
execution.
it causes a problems that calls processes directly under a root twice
(ref. the added test case)
it looks confusing behavior.

in this PR, i tried to fix the problem by just skipping creating a fresh
root for the first test execution

---------

Co-authored-by: Sam <sam@sksamuel.com>