Add line endings to unix option in assertions. (#5288)
Fixes #3161

Introduces StringPreprocessor's so we can add more string manipulation
in the future if needed.
Currently only covering StringEq and startsWith / endsWith, for
simplicity.

<!-- 
If this PR updates documentation, please update all relevant versions of
the docs, see:
https://github.com/kotest/kotest/tree/master/documentation/versioned_docs
The documentation at
https://github.com/kotest/kotest/tree/master/documentation/docs is the
documentation for the next minor or major version _TO BE RELEASED_
-->